orchestra-stage/app/requirements.txt
fastapi
uvicorn
docker
requests
python-dotenv
groq
jinja2
python-multipart